apply suggestions from Philipp review
